01-VanillaGAN/requirements.txt
tensorflow-gpu==1.1.0
numpy==1.12.1
scikit-learn==0.18.1
